requirements.txt
flask
flask-cors
rembg[cpu]
onnxruntime
Pillow